backend/backend/requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.1
Django==4.1
django-cors-headers==3.11.0
django-rest-framework==0.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
http-client==0.1.22
idna==3.3
pycurl==7.45.1
PyJWT==2.3.0
pytz==2022.2.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.12
